requirements.txt
Flask==3.0.2
pypdf==4.1.0
requests==2.31.0
PyYAML>=6.0.1,<7.0